Move Program and ProgramJob REST adapters to separate files
